requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic[email]==2.7.1
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
pytest==8.2.0
httpx==0.27.0
pytest-cov==5.0.0
pytest-playwright==0.5.0
playwright==1.44.0
